requirements.txt
whisper-timestamped==1.15.6
tk==0.1.0
pillow==11.0.0
numpy==2.0.2
moviepy==1.0.3
ffmpy==0.4.0
ffmpeg-python==0.2.0
ffmpeg==1.4
pandas==2.2.3